backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-socketio==5.11.4
python-multipart==0.0.20
langchain==0.3.11
langchain-openai==0.2.10
openai==1.57.2
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiohttp==3.11.10
